[TD-26] feat: Implement Lazy Loading (Homepage)
